+++ buzzid/detect/processing.py
@@ -6,7 +6,7 @@ import pandas as pd
 import matplotlib.pyplot as plt
 
 from buzzid.identify.tag_reader import process_tag_detection
-from buzzid.timestamp_reader import add_timestamps_to_dataframe
+from buzzid.timestamp_reader.timestamp_reader import add_timestamps_to_dataframe
 
 
 class PostProcessor:
